embodied_agent/requirements.txt
gymnasium>=1.2.1
matplotlib>=3.10.6
pybullet>=3.2.7
pyyaml>=6.0.2
stable-baselines3>=2.7.0